% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cisbpTFcat}
\alias{cisbpTFcat}
\title{cisbpTFcat: data.frame with information on CISBP TFs for human}
\format{data.frame}
\source{
\url{http://cisbp.ccbr.utoronto.ca/bulk.php} select Homo_sapiens
}
\usage{
cisbpTFcat
}
\description{
cisbpTFcat: data.frame with information on CISBP TFs for human
}
\note{
Extracted March 2018
}
\examples{
head(TFutils::cisbpTFcat)
}
\keyword{datasets}
