% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{encode690}
\alias{encode690}
\title{encode690: DataFrame extending AnnotationHub metadata about ENCODE cell line x TF ranges}
\format{DataFrame}
\source{
see metadata(encode690)
}
\usage{
encode690
}
\description{
encode690: DataFrame extending AnnotationHub metadata about ENCODE cell line x TF ranges
}
\examples{
names(TFutils::encode690)
TFutils::encode690[,1:5]
}
\keyword{datasets}
