% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fimo16}
\alias{fimo16}
\title{fimo16: GenomicFiles instance to AWS S3-resident FIMO bed for 16 TFs}
\format{GenomicFiles for a TabixFileList}
\source{
K. Glass FIMO runs, see \url{https://doi.org/10.1016/j.celrep.2017.10.001}
}
\usage{
fimo16
}
\description{
fimo16: GenomicFiles instance to AWS S3-resident FIMO bed for 16 TFs
}
\examples{
TFutils::fimo16
}
\keyword{datasets}
