% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fimoMap}
\alias{fimoMap}
\title{fimoMap: table with Mnnnn (motif PWM tags) and HGNC symbols for TFs}
\format{data.frame}
\source{
Kimberly Glass (rekrg@channing.harvard.edu)
}
\usage{
fimoMap
}
\description{
fimoMap: table with Mnnnn (motif PWM tags) and HGNC symbols for TFs
}
\examples{
head(TFutils::fimoMap)
}
\keyword{datasets}
