% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFCatalog-class.R
\name{HGNCmap}
\alias{HGNCmap}
\title{simple accessor for HGNCmap component of TFCatalog}
\usage{
HGNCmap(x)
}
\arguments{
\item{x}{instance of TFCatalog}
}
\value{
dataframe instance
}
\description{
simple accessor for HGNCmap component of TFCatalog
}
\examples{
HGNCmap
}
