% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hocomoco.mono}
\alias{hocomoco.mono}
\title{hocomoco.mono: data.frame with information on HOCOMOCO TFs for human}
\format{data.frame}
\source{
\url{http://hocomoco11.autosome.ru/human/mono?full=true}
}
\usage{
hocomoco.mono
}
\description{
hocomoco.mono: data.frame with information on HOCOMOCO TFs for human
}
\note{
Extracted March 2018
}
\examples{
head(TFutils::hocomoco.mono)
}
\keyword{datasets}
