% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfutils.R
\name{importFIMO_local_split}
\alias{importFIMO_local_split}
\title{utility to read FIMO outputs from local resource(cluster), assuming bed text split by chromosome}
\usage{
importFIMO_local_split(tf, chr)
}
\arguments{
\item{tf}{character(1) file id}

\item{chr}{character(1) chromosome name}
}
\value{
data.table instance
}
\description{
utility to read FIMO outputs from local resource(cluster), assuming bed text split by chromosome
}
\examples{
requireNamespace("GenomicRanges")
requireNamespace("IRanges")
importFIMO_local_split("M5946_1", "chr1")
dim(importFIMO_local_split("M5946_1", "chr17"))
}
