% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_tf}
\alias{metadata_tf}
\title{metadata_tf: list with metadata (motif_if and hgnc_symbol) about all the CISBP FIMO scan TF bed files}
\format{list}
\source{
K. Glass ran FIMO
}
\usage{
metadata_tf
}
\description{
metadata_tf: list with metadata (motif_if and hgnc_symbol) about all the CISBP FIMO scan TF bed files
}
\examples{
TFutils::metadata_tf
}
\keyword{datasets}
