% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{named_tf}
\alias{named_tf}
\title{named_tf: named list with the names being the hgnc_symbol of the motif_id}
\format{list}
\source{
K. Glass ran FIMO
}
\usage{
named_tf
}
\description{
named_tf: named list with the names being the hgnc_symbol of the motif_id
}
\examples{
TFutils::named_tf
named_tf[["VDR"]]
}
\keyword{datasets}
