% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFCatalog-class.R
\docType{methods}
\name{show,TFCatalog-method}
\alias{show,TFCatalog-method}
\title{produce a concise report on TFCatalog instance}
\usage{
\S4method{show}{TFCatalog}(object)
}
\arguments{
\item{object}{instance of TFCatalog}
}
\value{
side effect
}
\description{
produce a concise report on TFCatalog instance
}
