% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFCatalog-class.R
\docType{class}
\name{TFCatalog-class}
\alias{TFCatalog-class}
\alias{"TFCatalog-class"}
\title{define a structure to hold information about TFs from diverse reference sources}
\description{
define a structure to hold information about TFs from diverse reference sources
}
\section{Slots}{

\describe{
\item{\code{name}}{character}

\item{\code{nativeIds}}{character tokens used by the provider to enumerate transcription factors}

\item{\code{HGNCmap}}{data.frame with atleast two columns,
native id as first column and HGNC symbol as second column}

\item{\code{metadata}}{ANY}
}}

\note{
This class respects the notions that 1) a source of information
about transcription factors should have a name, 2) each source
has its own 'native' nomenclature for the factors themselves,
3) it is common to use the gene symbol to refer to the transctiption
factor, and 4) additional metadata will frequently be required to
establish information about provenance of assertions about transcription
factors.
}
