% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gob.R
\name{TFtargs}
\alias{TFtargs}
\title{gadget to help sort through tags naming TFs}
\usage{
TFtargs(gscoll = TFutils::tftColl, initTF = "VDR_Q3",
  gadtitle = "Search for a TF; its targets will be checked for mapped status in GWAS catalog")
}
\arguments{
\item{gscoll}{a GSEABase GeneSetCollection}

\item{initTF}{character(1) initial TF string for app}

\item{gadtitle}{character(1) a title for the gadget panel}
}
\value{
on app conclusion a data.frame is returned
}
\description{
gadget to help sort through tags naming TFs
}
\note{
Will use TFutils::gwascat_hg19_chr17 to look for 'MAPPED_GENE' field entries matching targets, also hardcoded to use org.Hs.eg.db to map symbols
}
\examples{
if (interactive()) TFtargs()
}
