% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tftCollMap}
\alias{tftCollMap}
\title{tftCollMap: data.frame with information on MSigDb TFs for human}
\format{data.frame}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/genesets.jsp?collection=TFT}
}
\usage{
tftCollMap
}
\description{
tftCollMap: data.frame with information on MSigDb TFs for human
}
\note{
Annotation of TFs is ad-hoc.  GeneSet names were tokenized, splitting
by underscore, and then fragments were matched to SYMBOL and ALIAS elements
of org.Hs.eg.db.  Extracted March 2018
}
\examples{
head(TFutils::tftCollMap)
}
\keyword{datasets}
