% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{URL_s3_tf}
\alias{URL_s3_tf}
\title{utility to generate link to biocfound bucket for FIMO TFBS scores}
\usage{
URL_s3_tf(tag = "M3433")
}
\arguments{
\item{tag}{character(1) token identifying TF, can be an HGNC gene name or Mnnnn PWM tag.
It must be findable in TFutils::fimoMap table.}
}
\value{
character(1) URL
}
\description{
utility to generate link to biocfound bucket for FIMO TFBS scores
}
\examples{
URL_s3_tf
}
